requirements.txt
cairocffi==1.6.1
CairoSVG==2.7.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.2.0
cssselect2==0.7.0
cycler==0.12.1
defusedxml==0.7.1
fonttools==4.45.1
idna==3.6
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.2
packaging==23.2
pandas==2.2.0
Pillow==10.1.0
pyarrow==15.0.0
pycparser==2.21
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
setuptools==69.0.3
six==1.16.0
threadpoolctl==3.2.0
tinycss2==1.2.1
tzdata==2023.4
urllib3==2.2.0
webencodings==0.5.1
